Thu, Jun 8, 2023, 7:17 PM +02:00

+++ volume/volume.c
@@ -34,10 +34,10 @@ int main(int argc, char *argv[])
     float factor = atof(argv[3]);
 
     // TODO: Copy header from input file to output file
-    int arr_header[HEADER_SIZE];
+    uint8_t arr_header[HEADER_SIZE];
     fread(arr_header, sizeof(uint8_t), HEADER_SIZE, input);
     fwrite(arr_header,sizeof(uint8_t), HEADER_SIZE, output);
-    
+
 
     // TODO: Read samples from input file and write updated data to output file
